eb-flask/requirements.txt
abstract-rendering==0.5.1
alabaster==0.7.6
anaconda-client==1.1.0
appnope==0.1.0
appscript==1.0.1
argcomplete==1.0.0
astropy==1.0.5
awscli==1.10.1
Babel==2.1.1
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.4.1
bitarray==0.8.1
blaze==0.8.3
bokeh==0.10.0
boto==2.38.0
botocore==1.3.23
Bottleneck==1.0.0
cdecimal==2.3
cffi==1.2.1
characteristic==14.3.0
clyent==0.4.0
colorama==0.3.3
conda==3.18.3
conda-build==1.18.1
conda-env==2.4.4
configobj==5.0.6
cryptography==1.0.2
cssselect==0.9.1
Cython==0.23.4
cytoolz==0.7.4
datashape==0.4.7
decorator==4.0.4
docutils==0.12
elasticsearch==2.3.0
enum34==1.0.4
fastcache==1.0.2
Flask==0.10.1
funcsigs==0.4
gevent==1.0.1
gevent-websocket==0.9.3
greenlet==0.4.9
grin==1.2.1
gunicorn==19.3.0
h5py==2.5.0
idna==2.0
ipaddress==1.0.14
ipykernel==4.1.1
ipython==4.0.0
ipython-genutils==0.1.0
ipywidgets==4.1.0
itsdangerous==0.24
jdcal==1.0
jedi==0.9.0
Jinja2==2.8
jmespath==0.9.0
jsonschema==2.4.0
jupyter==1.0.0
jupyter-client==4.1.1
jupyter-console==4.0.3
jupyter-core==4.0.6
llvmlite==0.7.0+3.g1ec568f
lxml==3.4.4
MarkupSafe==0.23
matplotlib==1.4.3
mistune==0.7.1
mock==1.3.0
multipledispatch==0.4.8
MySQL-python==1.2.5
nbconvert==4.0.0
nbformat==4.0.1
networkx==1.10
nltk==3.1
nose==1.3.7
notebook==4.0.6
numba==0.21.0
numexpr==2.4.4
numpy==1.10.1
oauthlib==1.0.3
odo==0.3.4
openpyxl==2.2.6
pandas==0.17.0
path.py==0.0.0
patsy==0.4.0
pbr==1.8.1
pep8==1.6.2
pexpect==3.3
pickleshare==0.5
Pillow==3.0.0
ply==3.8
psutil==3.2.2
ptyprocess==0.5
py==1.4.30
pyasn1==0.1.9
pyasn1-modules==0.0.8
PyAudio==0.2.7
pycosat==0.6.1
pycparser==2.14
pycrypto==2.6.1
pycurl==7.19.5.1
pyflakes==1.0.0
Pygments==2.0.2
pyOpenSSL==0.15.1
pyparsing==2.0.3
pytest==2.8.1
python-dateutil==2.4.2
pytz==2015.6
PyYAML==3.11
pyzmq==14.7.0
qtconsole==4.1.0
queuelib==1.4.2
redis==2.10.3
requests==2.8.1
requests-oauthlib==0.6.1
rope==0.9.4
rsa==3.3
scikit-image==0.11.3
scikit-learn==0.16.1
scipy==0.16.0
Scrapy==1.0.4
selenium==2.48.0
service-identity==14.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.0
sockjs-tornado==1.0.1
Sphinx==1.3.1
sphinx-rtd-theme==0.1.7
spyder==2.3.7
SQLAlchemy==1.0.9
statsmodels==0.6.1
sympy==0.7.6.1
tables==3.2.2
terminado==0.5
Theano==0.7.0
toolz==0.7.4
tornado==4.2.1
traitlets==4.0.0
tweepy==3.5.0
Twisted==15.5.0
ujson==1.33
unicodecsv==0.14.1
urllib3==1.14
w3lib==1.13.0
Werkzeug==0.10.4
wheel==0.26.0
xlrd==0.9.4
XlsxWriter==0.7.7
xlwings==0.4.1
xlwt==1.0.0
zope.interface==4.1.3
